Fixing non-determinism and changing names
The usage of several updates on one transition in 'step_behavior' causes non-determinism in the aotomata. It is probably some optimization in Supremica that removes the updates if it is a tautology. Now it is changed to three transitions, one for each separate update.

The marking of 'step_behavior' is changed to only the initial state. The 'possible_inputs' event is simultaneously changed to being uncontrollable. In this way a supervisor cannot halt execution. It has to allow progression to the marked state, and from there an uncontrollable event will continue a new cycle.

Shortening the road to 30 cells in order to ease the load on the compiler. It is a conjecture that this will not affect the behavior of a supervisor, since 30 cells are still enough to separate the two vehicles considerably.

Changing all names with the substring 'adv' to to the substring 'ego' to improve readability.