Updating guards for safe distance to account for wrap around
Changing the guards for maintaining safe distance when driving forward. The previous guards did not work when one vehicle was close to the start or end of the road. In those places the safe positions form a band, limited from below and above (since the occupied positions would wrap around and be placed both in the beginning and the end of the road).