Changing to Mealy machine and removing unnecessary specifications
Changing to Mealy machine to get the example working without assuming that the env_veh cannot be stationary forever. This feels more realistic, and is also possible to model with Supremica. Mealy's output is dependent on state and input (contrary to Moore which only is dependent on state). Somehow, this makes it possible to predict future moves better.
As a consequence of this change, the specifications prohibiting lane changes when the vehicles are next to each other can be omitted. This would anyway be prohibited by the specifications saying that the vehicles may not be dangerously close when they are in the same lane.